pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.99.1"}
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pydantic = {version = "^1.10.4", extras = ["dotenv"]}
yarl = "^1.8.2"
ujson = "^5.7.0"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
asyncpg = "^0.27.0"
redis = {version = "^4.4.2", extras = ["hiredis"]}
httptools = "^0.5.0"
prometheus-client = "^0.16.0"
prometheus-fastapi-instrumentator = "5.9.1"
sentry-sdk = "^1.14.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-exporter-otlp = "^1.15.0"
opentelemetry-instrumentation = "^0.36b0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-logging = "^0.36b0"
opentelemetry-instrumentation-redis = "^0.36b0"
opentelemetry-instrumentation-asyncpg = "^0.36b0"
rocketry = "^2.5.1"
talipp = "^1.9.1"
gevent = "^22.10.2"
signalr-client = "^0.0.7"
python-socketio = "^5.8.0"
websocket-client = "^1.6.1"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiogram = "^2.25.1"
loguru = "^0.7.0"
darglint = "^1.8.1"
pandas = "^2.0.3"
joblib = "^1.3.2"
scikit-learn = "^1.3.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.6"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["server",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "SERVER_ENVIRONMENT=pytest",
    "SERVER_DB_BASE=server_test",
    "SERVER_SENTRY_DSN=",
]
[tool.aerich]
tortoise_orm = "server.db.config.TORTOISE_CONFIG"
location = "./server/db/migrations"
src_folder = "./server"

[fastapi-template.options]
project_name = "server"
api_type = "rest"
enable_redis = "True"
enable_rmq = "None"
ci_type = "none"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "server"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "True"
add_dummy = "True"
orm = "tortoise"
self_hosted_swagger = "None"
prometheus_enabled = "True"
sentry_enabled = "True"
otlp_enabled = "True"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
